+++ noodles-gff/src/record_buf/attributes.rs
@@ -2,17 +2,11 @@
 
 pub mod field;
 
-use std::{
-    error, fmt,
-    ops::{Deref, DerefMut},
-    str::{self, FromStr},
-};
+use std::ops::{Deref, DerefMut};
 
 use indexmap::IndexMap;
 
 use self::field::{Tag, Value};
-
-const DELIMITER: char = ';';
 
 /// GFF record attributes.
 ///
@@ -35,37 +29,6 @@ impl DerefMut for Attributes {
     }
 }
 
-/// An error returned when raw attributes fail to parse.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum ParseError {
-    /// A field is invalid.
-    InvalidField(field::ParseError),
-}
-
-impl error::Error for ParseError {
-    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
-        match self {
-            Self::InvalidField(e) => Some(e),
-        }
-    }
-}
-
-impl fmt::Display for ParseError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::InvalidField(e) => {
-                write!(f, "invalid field")?;
-
-                if let Some(key) = e.key() {
-                    write!(f, ": {key}")?;
-                }
-
-                Ok(())
-            }
-        }
-    }
-}
-
 impl Extend<(Tag, Value)> for Attributes {
     fn extend<T: IntoIterator<Item = (Tag, Value)>>(&mut self, iter: T) {
         self.0.extend(iter);
@@ -77,70 +40,5 @@ impl FromIterator<(Tag, Value)> for Attributes {
         let mut attributes = Self::default();
         attributes.extend(iter);
         attributes
-    }
-}
-
-impl FromStr for Attributes {
-    type Err = ParseError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        use self::field::parse_field;
-
-        if s.is_empty() {
-            return Ok(Self::default());
-        }
-
-        let mut map: IndexMap<Tag, Value> = IndexMap::new();
-
-        for raw_field in s.split(DELIMITER) {
-            let (key, value) = parse_field(raw_field).map_err(ParseError::InvalidField)?;
-
-            map.entry(key)
-                .and_modify(|v| v.extend(value.iter().cloned()))
-                .or_insert(value);
-        }
-
-        Ok(Self(map))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_from_str() -> Result<(), ParseError> {
-        let actual: Attributes = "".parse()?;
-        let expected = Attributes::default();
-        assert_eq!(actual, expected);
-
-        let s = "gene_id=ndls0";
-        let actual: Attributes = s.parse()?;
-        let expected = [(Tag::from("gene_id"), Value::from("ndls0"))]
-            .into_iter()
-            .collect();
-        assert_eq!(actual, expected);
-
-        let s = "gene_id=ndls0;gene_name=gene0";
-        let actual: Attributes = s.parse()?;
-        let expected = [
-            (Tag::from("gene_id"), Value::from("ndls0")),
-            (Tag::from("gene_name"), Value::from("gene0")),
-        ]
-        .into_iter()
-        .collect();
-        assert_eq!(actual, expected);
-
-        let s = "gene_id=ndls0;gene_id=ndls1";
-        let actual: Attributes = s.parse()?;
-        let expected = [(
-            Tag::from("gene_id"),
-            Value::from(vec![String::from("ndls0"), String::from("ndls1")]),
-        )]
-        .into_iter()
-        .collect();
-        assert_eq!(actual, expected);
-
-        Ok(())
     }
 }